backend/requirements.txt
pyramid==2.0
pyramid-cors==0.3
waitress==2.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
google-generativeai==0.3.2
requests==2.31.0